+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @openfn/cli
 
+## 1.17.1
+
+### Patch Changes
+
+- a0a1cb7: Add log option for checkout and merge
+
+## 1.17.0
+
+### Minor Changes
+
+- New command: openfn project version
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/project@0.6.0
+
+## 1.16.2
+
+### Patch Changes
+
+- 8a50703: Allow a project to be checked out from a direct path to a project file
+- 7d16875: When merging, allow the source project to specified as a path to a project file
+- Updated dependencies [81b97c3]
+- Updated dependencies [43be979]
+- Updated dependencies [1f8d65e]
+- Updated dependencies [25c7a2b]
+- Updated dependencies [04a89e2]
+- Updated dependencies [7d16875]
+  - @openfn/project@0.5.1
+  - @openfn/lexicon@1.2.4
+
 ## 1.16.1
 
 ### Patch Changes